pyproject.toml
[project]
name = "test-project-ivi"
version = "0.1.0"
description = ""
authors = [
    {name = "Max_Razzhivin",email = "max.nvo06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "selene (==2.0.0rc9)",
    "allure-pytest (>=2.14.2,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "appium-python-client (==5.1.1)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
    ]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
